+++ packages/aws-amplify-angular/src/assets/default-sign-up-fields.ts
@@ -1,0 +1,28 @@
+export default [
+  {
+    label: 'Username',
+    key: 'username',
+    required: false,
+    displayOrder: 1
+  },
+  {
+    label: 'Password',
+    key: 'password',
+    required: true,
+    type: 'password',
+    displayOrder: 2,
+  },
+  {
+    label: 'Email',
+    key: 'email',
+    required: true,
+    type: 'email',
+    displayOrder: 3
+  },
+  {
+    label: 'Phone Number',
+    key: 'phone_number',
+    required: true,
+    displayOrder: 4
+  }
+];