+++ packages/aws-amplify-angular/src/components/authenticator/sign-up-component/sign-up.factory.ts
@@ -1,9 +1,15 @@
-import { Component, Input, OnInit, ViewChild, ComponentFactoryResolver, OnDestroy } from '@angular/core';
-
+import { 
+  Component,
+  Input,
+  OnInit,
+  ViewChild,
+  ComponentFactoryResolver,
+  OnDestroy 
+} from '@angular/core';
 import { DynamicComponentDirective } from '../../../directives/dynamic.component.directive';
 import { ComponentMount }      from '../../component.mount';
 import { SignUpClass } from './sign-up.class';
-import { SignUpComponentIonic } from './sign-up.component.ionic'
+import { SignUpComponentIonic } from './sign-up.component.ionic';
 import { SignUpComponentCore } from './sign-up.component.core';
 import { AuthState } from '../../../providers';
 
@@ -18,6 +24,7 @@ import { AuthState } from '../../../providers';
 export class SignUpComponent implements OnInit, OnDestroy {
   @Input() framework: string;
   @Input() authState: AuthState;
+  @Input() signUpConfig: any;
   @ViewChild(DynamicComponentDirective) componentHost: DynamicComponentDirective;
 
   constructor(private componentFactoryResolver: ComponentFactoryResolver) { }
@@ -30,14 +37,23 @@ export class SignUpComponent implements OnInit, OnDestroy {
 
   loadComponent() {
 
-    let authComponent = this.framework && this.framework.toLowerCase() === 'ionic' ? new ComponentMount(SignUpComponentIonic,{authState: this.authState}) : new ComponentMount(SignUpComponentCore, {authState: this.authState});
+    const authComponent = this.framework && this.framework.toLowerCase() === 'ionic' ?
+    new ComponentMount(SignUpComponentIonic, {
+      authState: this.authState,
+      signUpConfig: this.signUpConfig
+    }) :
+    new ComponentMount(SignUpComponentCore, {
+      authState: this.authState,
+      signUpConfig: this.signUpConfig
+    });
 
-    let componentFactory = this.componentFactoryResolver.resolveComponentFactory(authComponent.component);
+    const componentFactory = this.componentFactoryResolver
+    .resolveComponentFactory(authComponent.component);
 
-    let viewContainerRef = this.componentHost.viewContainerRef;
+    const viewContainerRef = this.componentHost.viewContainerRef;
     viewContainerRef.clear();
 
-    let componentRef = viewContainerRef.createComponent(componentFactory);
+    const componentRef = viewContainerRef.createComponent(componentFactory);
     (<SignUpClass>componentRef.instance).data = authComponent.data;
   }
 }