+++ packages/aws-amplify-angular/src/providers/amplify.service.ts
@@ -24,7 +24,7 @@ export class AmplifyService {
   private _cache: any;
   private _pubsub: any;
   private _interactions: InteractionsClass;
-
+  private _logger: any; 
   private _authState = new Subject<AuthState>();
   authStateChange$ = this._authState.asObservable();
 
@@ -38,7 +38,7 @@ export class AmplifyService {
     this._cache = Amplify.Cache;
     this._pubsub = Amplify.PubSub;
     this._interactions = Amplify.Interactions;
-
+    this._logger =  Amplify.Logger;
   }
 
   auth(): AuthClass { return this._auth; }
@@ -48,6 +48,7 @@ export class AmplifyService {
   interactions(): InteractionsClass { return this._interactions; }
   cache(): any { return this._cache; }
   pubsub(): any { return this._pubsub; }
+  logger(name, level): Logger { return new this._logger(name, level); }
 
   authState() { return this._authState; }
   setAuthState(state: AuthState) { this._authState.next(state); }