+++ packages/aws-amplify-react-native/src/Auth/SignUp.js
@@ -18,7 +18,8 @@ import {
     TextInput, 
     Button, 
     TouchableWithoutFeedback,
-    Keyboard
+    Keyboard,
+    Picker
 } from 'react-native';
 import {
     Auth,
@@ -33,6 +34,8 @@ import {
     AmplifyButton
 } from '../AmplifyUI';
 import AuthPiece from './AuthPiece';
+import defaultSignUpFields from './common/default-sign-in-fields'
+
 
 const logger = new Logger('SignUp');
 export default class SignUp extends AuthPiece {
@@ -40,64 +43,170 @@ export default class SignUp extends AuthPiece {
         super(props);
 
         this._validAuthStates = ['signUp'];
-        this.state = {
-            username: null,
-            password: null,
-            email: null,
-            phone_number: null
-        }
-
+        this.state = {};
         this.signUp = this.signUp.bind(this);
+        this.sortFields = this.sortFields.bind(this);
+        this.getDefaultDialCode = this.getDefaultDialCode.bind(this);
+        this.checkCustomSignUpFields = this.checkCustomSignUpFields.bind(this);
+        this.defaultSignUpFields = defaultSignUpFields;
+        this.needPrefix = this.needPrefix.bind(this);
+    }
+
+    validate() {
+        const invalids = [];
+        this.signUpFields.map((el) => {
+            if (el.required && !this.state[el.key]) {
+                el.invalid = true;
+                invalids.push(el.label);
+            } else {
+                el.invalid = false;
+            }        
+        });
+        return invalids;
+      }
+
+    sortFields() {
+        if (this.checkCustomSignUpFields()) {
+
+          if (!this.props.signUpConfig || !this.props.signUpConfig.hideDefaults) {
+            // see if fields passed to component should override defaults
+            this.defaultSignUpFields.forEach((f, i) => {
+              const matchKey = this.signUpFields.findIndex((d) => {
+                return d.key === f.key;
+              });
+              if (matchKey === -1) {
+                this.signUpFields.push(f);
+              }
+            });
+          }
+    
+          /* 
+            sort fields based on following rules:
+            1. Fields with displayOrder are sorted before those without displayOrder
+            2. Fields with conflicting displayOrder are sorted alphabetically by key
+            3. Fields without displayOrder are sorted alphabetically by key
+          */
+          this.signUpFields.sort((a, b) => {
+            if (a.displayOrder && b.displayOrder) {
+              if (a.displayOrder < b.displayOrder) {
+                return -1;
+              } else if (a.displayOrder > b.displayOrder) {
+                return 1;
+              } else {
+                if (a.key < b.key) {
+                  return -1;
+                } else {
+                  return 1;
+                }
+              }
+            } else if (!a.displayOrder && b.displayOrder) {
+              return 1;
+            } else if (a.displayOrder && !b.displayOrder) {
+              return -1;
+            } else if (!a.displayOrder && !b.displayOrder) {
+              if (a.key < b.key) {
+                return -1;
+              } else {
+                return 1;
+              }
+            }
+          });
+        } else {
+          this.signUpFields = this.defaultSignUpFields;
+        }
+    }
+
+    needPrefix(key) {
+        const field = this.signUpFields.find(e => e.key === key);
+        if (key.indexOf('custom:') !== 0) {
+          return field.custom ;
+        } else if (key.indexOf('custom:') === 0 && field.custom === false) {
+          logger.warn('Custom prefix prepended to key but custom field flag is set to false');
+        }
+        return null;
+    }
+
+
+    getDefaultDialCode() {
+        return this.props.signUpConfig &&
+        this.props.signUpConfig.defaultCountryCode  &&
+        countryDialCodes.indexOf(`+${this.props.signUpConfig.defaultCountryCode}`) !== '-1' ?
+        `+${this.props.signUpConfig.defaultCountryCode}` :
+        "+1"
+    }
+
+    checkCustomSignUpFields() {
+        return this.props.signUpConfig &&
+        this.props.signUpConfig.signUpFields &&
+        this.props.signUpConfig.signUpFields.length > 0
     }
 
     signUp() {
-        const { username, password, email, phone_number } = this.state;
-        logger.debug('Sign Up for ' + username);
-        Auth.signUp(username, password, email, phone_number)
-            .then(data => {
-                logger.debug(data);
-                this.changeState('confirmSignUp', username);
-            })
-            .catch(err => this.error(err));
+        const validation = this.validate();
+        if (validation && validation.length > 0) {
+          return this.error(`The following fields need to be filled out: ${validation.join(', ')}`);
+        }
+        if (!Auth || typeof Auth.signUp !== 'function') {
+            throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
+        }
+
+        let signup_info = {
+            username: this.state.username,
+            password: this.state.password,
+            attributes: {
+                
+            }
+        };
+
+        const inputKeys = Object.keys(this.state);
+        const inputVals = Object.values(this.state);
+
+        inputKeys.forEach((key, index) => {
+            if (!['username', 'password', 'checkedValue'].includes(key)) {
+                const newKey = `${this.needPrefix(key) ? 'custom:' : ''}${key}`;
+                signup_info.attributes[newKey] = inputVals[index];
+            }
+        });
+
+        console.log('signup_info', signup_info)
+
+        Auth.signUp(signup_info).then((data) => {
+            this.changeState('confirmSignUp', data.user.username)
+        })
+        .catch(err => this.error(err));
     }
 
     showComponent(theme) {
+        if (this.checkCustomSignUpFields()) {
+            this.signUpFields = this.props.signUpConfig.signUpFields;
+        }
+        this.sortFields();
         return (
             <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
                 <View style={theme.section}>
                     <Header theme={theme}>{I18n.get('Create a new account')}</Header>
                     <View style={theme.sectionBody}>
-                        <FormField
-                            theme={theme}
-                            onChangeText={(text) => this.setState({ username: text })}
-                            label={I18n.get('Username')}
-                            placeholder={I18n.get('Enter your username')}
-                            required={true}
-                        />
-                        <FormField
-                            theme={theme}
-                            onChangeText={(text) => this.setState({ password: text })}
-                            label={I18n.get('Password')}
-                            placeholder={I18n.get('Enter your password')}
-                            secureTextEntry={true}
-                            required={true}
-                        />
-                        <FormField
-                            theme={theme}
-                            onChangeText={(text) => this.setState({ email: text })}
-                            label={I18n.get('Email')}
-                            placeholder={I18n.get('Enter your email')}
-                            keyboardType="email-address"
-                            required={true}
-                        />
-                        <FormField
-                            theme={theme}
-                            onChangeText={(text) => this.setState({ phone_number: text })}
-                            label={I18n.get('Phone Number')}
-                            placeholder={I18n.get('Enter your phone number')}
-                            keyboardType="phone-pad"
-                            required={true}
-                        />
+                    {
+                        this.signUpFields.map((field) => {
+                            return  (
+                                <FormField
+                                    key = {field.key}
+                                    theme={theme}
+                                    type={field.type}
+                                    secureTextEntry={field.type === 'password' ? true: false}
+                                    onChangeText={(text) => {
+                                            const stateObj = this.state;
+                                            stateObj[field.key] = text;
+                                            this.setState(stateObj)
+                                        }
+                                    }
+                                    label={I18n.get(field.label)}
+                                    placeholder={I18n.get(field.placeholder)}
+                                    required={field.required}
+                                />
+                            )
+                        })
+                    }
                         <AmplifyButton
                             text={I18n.get('Sign Up').toUpperCase()}
                             theme={theme}
@@ -118,4 +227,6 @@ export default class SignUp extends AuthPiece {
             </TouchableWithoutFeedback>
         );
     }
+
+
 }