+++ packages/aws-amplify-react/src/Auth/SignUp.jsx
@@ -33,103 +33,220 @@ import {
 } from '../Amplify-UI/Amplify-UI-Components-React';
 
 import countryDialCodes from './common/country-dial-codes.js';
+import defaultSignUpFields from './common/default-sign-in-fields'
+import { valid } from 'semver';
+
+
 export default class SignUp extends AuthPiece {
     constructor(props) {
         super(props);
 
         this._validAuthStates = ['signUp'];
         this.signUp = this.signUp.bind(this);
+        this.sortFields = this.sortFields.bind(this);
+        this.getDefaultDialCode = this.getDefaultDialCode.bind(this);
+        this.checkCustomSignUpFields = this.checkCustomSignUpFields.bind(this);
+        this.defaultSignUpFields = defaultSignUpFields;
+        this.needPrefix = this.needPrefix.bind(this);
+    }
 
-        this.inputs = {
-            dial_code: "+1",
-        };
+    validate() {
+        const invalids = [];
+        this.signUpFields.map((el) => {
+          if (el.key !== 'phone_number') {
+            if (el.required && !this.inputs[el.key]) {
+              el.invalid = true;
+              invalids.push(el.label);
+            } else {
+              el.invalid = false;
+            }        
+          } else {
+            if (el.required && (!this.inputs.dial_code || !this.inputs.phone_line_number)) {
+              el.invalid = true;
+              invalids.push(el.label);
+            } else {
+              el.invalid = false;
+            }
+          }
+        });
+        return invalids;
+    }
+
+    sortFields() {
+        if (this.checkCustomSignUpFields()) {
+
+          if (!this.props.signUpConfig || !this.props.signUpConfig.hideDefaults) {
+            // see if fields passed to component should override defaults
+            this.defaultSignUpFields.forEach((f, i) => {
+              const matchKey = this.signUpFields.findIndex((d) => {
+                return d.key === f.key;
+              });
+              if (matchKey === -1) {
+                this.signUpFields.push(f);
+              }
+            });
+          }
+    
+          /* 
+            sort fields based on following rules:
+            1. Fields with displayOrder are sorted before those without displayOrder
+            2. Fields with conflicting displayOrder are sorted alphabetically by key
+            3. Fields without displayOrder are sorted alphabetically by key
+          */
+          this.signUpFields.sort((a, b) => {
+            if (a.displayOrder && b.displayOrder) {
+              if (a.displayOrder < b.displayOrder) {
+                return -1;
+              } else if (a.displayOrder > b.displayOrder) {
+                return 1;
+              } else {
+                if (a.key < b.key) {
+                  return -1;
+                } else {
+                  return 1;
+                }
+              }
+            } else if (!a.displayOrder && b.displayOrder) {
+              return 1;
+            } else if (a.displayOrder && !b.displayOrder) {
+              return -1;
+            } else if (!a.displayOrder && !b.displayOrder) {
+              if (a.key < b.key) {
+                return -1;
+              } else {
+                return 1;
+              }
+            }
+          });
+        } else {
+          this.signUpFields = this.defaultSignUpFields;
+        }
+    }
+
+    needPrefix(key) {
+        const field = this.signUpFields.find(e => e.key === key);
+        if (key.indexOf('custom:') !== 0) {
+          return field.custom ;
+        } else if (key.indexOf('custom:') === 0 && field.custom === false) {
+            logger.warn('Custom prefix prepended to key but custom field flag is set to false');
+          
+        }
+        return null;
+    }
+
+    getDefaultDialCode() {
+        return this.props.signUpConfig &&
+        this.props.signUpConfig.defaultCountryCode  &&
+        countryDialCodes.indexOf(`+${this.props.signUpConfig.defaultCountryCode}`) !== '-1' ?
+        `+${this.props.signUpConfig.defaultCountryCode}` :
+        "+1"
+    }
+
+    checkCustomSignUpFields() {
+        return this.props.signUpConfig &&
+        this.props.signUpConfig.signUpFields &&
+        this.props.signUpConfig.signUpFields.length > 0
     }
 
     signUp() {
-        const { username, password, email, dial_code='+1', phone_line_number } = this.inputs;
+        if (!this.inputs.dial_code) {
+            this.inputs.dial_code = this.getDefaultDialCode();
+        }
+        const validation = this.validate();
+        if (validation && validation.length > 0) {
+          return this.error(`The following fields need to be filled out: ${validation.join(', ')}`);
+        }
         if (!Auth || typeof Auth.signUp !== 'function') {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
         }
 
         let signup_info = {
-            username,
-            password, 
+            username: this.inputs.username,
+            password: this.inputs.password,
             attributes: {
-                email
+                
             }
         };
 
-        let phone_number = phone_line_number? `${dial_code}${phone_line_number.replace(/[-()]/g, '')}`: null;
+        const inputKeys = Object.keys(this.inputs);
+        const inputVals = Object.values(this.inputs);
 
-        if (phone_number) {
-            signup_info.attributes.phone_number = phone_number;
-        }
+        inputKeys.forEach((key, index) => {
+            if (!['username', 'password', 'checkedValue'].includes(key)) {
+              if (key !== 'phone_line_number' && key !== 'dial_code') {
+                const newKey = `${this.needPrefix(key) ? 'custom:' : ''}${key}`;
+                signup_info.attributes[newKey] = inputVals[index];
+              } else {
+                  signup_info.attributes['phone_number'] = `+${this.inputs.dial_code}${this.inputs.phone_line_number.replace(/[-()]/g, '')}`
+              }
+            }
+        });
 
-        Auth.signUp(signup_info).then(() => this.changeState('confirmSignUp', username))
+        Auth.signUp(signup_info).then((data) => {
+            this.changeState('confirmSignUp', data.user.username)
+        })
         .catch(err => this.error(err));
     }
 
     showComponent(theme) {
         const { hide } = this.props;
         if (hide && hide.includes(SignUp)) { return null; }
-
-        // console.log(countryDialCodes.sort());
-
+        if (this.checkCustomSignUpFields()) {
+            this.signUpFields = this.props.signUpConfig.signUpFields;
+        }
+        this.sortFields();
         return (
             <FormSection theme={theme}>
                 <SectionHeader theme={theme}>{I18n.get('Create a new account')}</SectionHeader>
                 <SectionBody theme={theme}>
-                    <FormField theme={theme}>
-                        <InputLabel>{I18n.get('Username')} *</InputLabel>
-                        <Input
-                            autoFocus
-                            placeholder={I18n.get('Create a username')}
-                            theme={theme}
-                            key="username"
-                            name="username"
-                            onChange={this.handleInputChange}
-                        />
-                    </FormField>
-                    <FormField theme={theme}>
-                        <InputLabel>{I18n.get('Password')} *</InputLabel>
-                        <Input
-                            placeholder={I18n.get('Create a password')}
-                            theme={theme}
-                            type="password"
-                            key="password"
-                            name="password"
-                            onChange={this.handleInputChange}
-                        />
-                    </FormField>
-                    <FormField theme={theme}>
-                        <InputLabel>{I18n.get('Email Address')} *</InputLabel>
-                        <Input
-                            placeholder="janedoe@email.com"
-                            theme={theme}
-                            key="email"
-                            name="email"
-                            onChange={this.handleInputChange}
-                        />
-                    </FormField>
-                    <FormField theme={theme}>
-                        <InputLabel>{I18n.get('Phone Number')}</InputLabel>
-                        <SelectInput theme={theme}>
-                            <select key="dial_code" name="dial_code" defaultValue="+1" onChange={this.handleInputChange}>
-                                {countryDialCodes.map(dialCode =>
-                                    <option key={dialCode} value={dialCode}>
-                                        {dialCode}
-                                    </option>
-                                )}
-                            </select>
-                            <Input
-                                placeholder="555-555-1212"
-                                theme={theme}
-                                key="phone_line_number"
-                                name="phone_line_number"
-                                onChange={this.handleInputChange}
-                            />
-                        </SelectInput>
-                    </FormField>
+                    {
+                        this.signUpFields.map((field) => {
+                            return field.key !== 'phone_number' ? (
+                                <FormField theme={theme} key={field.key}>
+                                {
+                                    field.required ? 
+                                    <InputLabel>{I18n.get(field.label)} *</InputLabel> :
+                                    <InputLabel>{I18n.get(field.label)}</InputLabel>
+                                }
+                                    <Input
+                                        autoFocus={
+                                            this.signUpFields.findIndex((f) => {
+                                                return f.key === field.key
+                                            }) === 0 ? true : false
+                                        }
+                                        placeholder={I18n.get(field.placeholder)}
+                                        theme={theme}
+                                        type={field.type}
+                                        name={field.key}
+                                        onChange={this.handleInputChange}
+                                    />
+                                </FormField>
+                            ) : (
+                                <FormField theme={theme} key="phone_number">
+                                    <InputLabel>{I18n.get('Phone Number')}</InputLabel>
+                                    <SelectInput theme={theme}>
+                                        <select name="dial_code" defaultValue={this.getDefaultDialCode()} 
+                                        onChange={this.handleInputChange}>
+                                            {countryDialCodes.map(dialCode =>
+                                                <option key={dialCode} value={dialCode}>
+                                                    {dialCode}
+                                                </option>
+                                            )}
+                                        </select>
+                                        <Input
+                                            placeholder={I18n.get(field.placeholder)}
+                                            theme={theme}
+                                            type="tel"
+                                            id="phone_line_number"
+                                            key="phone_line_number"
+                                            name="phone_line_number"
+                                            onChange={this.handleInputChange}
+                                        />
+                                    </SelectInput>
+                                </FormField>
+                            )
+                        })
+                    }
                 </SectionBody>
                 <SectionFooter theme={theme}>
                     <SectionFooterPrimaryContent theme={theme}>
@@ -147,4 +264,6 @@ export default class SignUp extends AuthPiece {
             </FormSection>
         );
     }
+
+
 }