+++ packages/aws-amplify-angular/CHANGELOG.md
@@ -3,16 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-<a name="2.0.12-beta.1"></a>
-## [2.0.12-beta.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.12-beta.0...aws-amplify-angular@2.0.12-beta.1) (2018-10-30)
+<a name="2.0.12-unstable.3"></a>
+## [2.0.12-unstable.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.12-unstable.2...aws-amplify-angular@2.0.12-unstable.3) (2018-10-30)
 
 
 
 
 **Note:** Version bump only for package aws-amplify-angular
 
-<a name="2.0.12-beta.0"></a>
-## [2.0.12-beta.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.11...aws-amplify-angular@2.0.12-beta.0) (2018-10-30)
+<a name="2.0.12-unstable.2"></a>
+## [2.0.12-unstable.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.12-unstable.1...aws-amplify-angular@2.0.12-unstable.2) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.12-unstable.1"></a>
+## [2.0.12-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.12-unstable.0...aws-amplify-angular@2.0.12-unstable.1) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-angular
+
+<a name="2.0.12-unstable.0"></a>
+## [2.0.12-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-angular@2.0.11...aws-amplify-angular@2.0.12-unstable.0) (2018-10-30)